# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Beyond The Label"
host: "localhost"
basePath: "/btl"
schemes:
- "https"
- "http"
paths:
  /customer:
    post:
      tags:
      - "Customer"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Customer create request"
        required: true
        schema:
          type: object
          required:
          - "name"
          - "logo_file_path"
          properties:
            name:
              type: string
              format: string
            logo_file_path:
              type: string
              format: binary
            intent_file_path:
              type: string
              format: binary
      responses:
        "200":
          description: "Customer created successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                format: string
              data:
                type: object
                properties:
                  customer_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    format: string
                  logo_file_path:
                    type: string
                  intent_file_path:
                    type: string
                  status:
                    type: boolean
                  training_status:
                    type: integer
                    format: int64
                  bucket_name:
                    type: string
                    format: string
                  is_deleted:
                    type: boolean
                  customer_display_id:
                    type: string
                    format: string
        "405":
          description: "Invalid input"
    put:
      tags:
      - "Customer"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Customer update request"
        required: true
        schema:
          type: object
          required:
          - "customer_id"
          properties:
            customer_id:
              type: string
              format: uuid
            logo_file_path:
              type: string
              format: binary
            intent_file_path:
              type: string
              format: binary
      responses:
        "200":
          description: "Customer updated successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                format: string
              data:
                type: object
                properties:
                  customer_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    format: string
                  logo_file_path:
                    type: string
                  intent_file_path:
                    type: string
                  status:
                    type: boolean
                  training_status:
                    type: integer
                    format: int64
                  bucket_name:
                    type: string
                    format: string
                  is_deleted:
                    type: boolean
                  customer_display_id:
                    type: string
                    format: string
        "405":
          description: "Invalid input"
    get:
      tags:
      - "Customer"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      responses:
        "405":
          description: "Invalid Request"
        "200":
          description: "Customer Details"
          schema:
            type: array
            items:
              type: object
              properties:
                customer_id:
                  type: string
                  format: uuid
                name:
                  type: string
                  format: string
                logo_file_path:
                  type: string
                intent_file_path:
                  type: string
                status:
                  type: boolean
                training_status:
                  type: integer
                  format: int64
                bucket_name:
                  type: string
                  format: string
                is_deleted:
                  type: boolean
                customer_display_id:
                  type: string
                  format: string
  /customer/{id}:
    delete:
      tags:
      - "Customer"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Customer id to delete"
        required: true
        type: string
        format: uuid
      responses:
        "200":
          description: "Customer deleted successfully"
        "405":
          description: "Invalid input"
  /customer/status:
    put:
      tags:
      - "Customer"
      consumes:
      - "application/octet-stream"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Customer status update request"
        required: true
        schema:
          type: object
          properties:
            current_customer_id:
              type: string
              format: uuid
            new_customer_id:
              type: string
              format: uuid
      responses:
        "200":
          description: "Customer status changed successfully"
        "405":
          description: "Invalid input"
  /customer/active:
    get:
      tags:
      - "Customer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "405":
          description: "Invalid Request"
        "200":
          description: "Customer Details"
          schema:
            type: object
            properties:
              customer_id:
                type: string
                format: uuid
              name:
                type: string
                format: string
              logo_file_path:
                type: string
              intent_file_path:
                type: string
              status:
                type: boolean
              training_status:
                type: integer
                format: int64
              bucket_name:
                type: string
                format: string
              is_deleted:
                type: boolean
              customer_display_id:
                type: string
                format: string
...
  /customer/{id}/intent:
    get:
      summary: Returns all intents for a specific customer
      description: This can be used to display all intents in a list which can then be cycled through to view particular intents. Filter intents by customer ID using CSV.
      tags:
      - Customer
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Customer id for intents"
        required: true
        type: string
      responses:
        "200":
          description: "All intents for this customer"
          schema:
            type: array
            items:
              $ref: '#/definitions/Intent'
        "400":
          description: ""
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                description: >
                  error:
                   * `First Error` - No Agent found for this customer!
                   * `Second Error` - No intents associated with this Agent ID!
                   * `Third Error` - There was an error while fetching an intent from DialogFlow!
                   * `Fourth Error` - There was an error while building intents response!
                enum:
                - "No Agent found for this customer!"
                - "No intents associated with this Agent ID!"
                - "There was an error while fetching an intent from DialogFlow!"
                - "There was an error while building intents response!"
                default: "No Agent found for this customer! No intents associated with this Agent ID! There was an error while fetching an intent from DialogFlow! There was an error while building intents response!"
    post:
      summary: Creates/Updates/Deletes intents mentioned in the CSV
      description: This will be used every time the CSV is updated. It'll create/update/delete intents mentioned in the CSV depending on the Action column
      tags:
      - Customer
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Customer id for intents"
        required: true
        type: string
      responses:
        "200":
          description: "Customer intents updated successfully."
        "400":
          description: ""
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                description: >
                  error:
                   * `First Error` - No Agent found for this customer!
                   * `Second Error` - No Intents file found at the path!
                   * `Third Error` - Customer intents update failed!
                enum:
                - "No Agent found for this customer!"
                - "No Intents file found at the path!"
                - "Customer intents update failed!"
                default: "No Agent found for this customer! No Intents file found at the path! Customer intents update failed!"
  /customer/{id}/intent/download:
    get:
      summary: Downloads a CSV file of customer intents
      description: Calls the get intent API, writes all the intents to the csv file and prompts the user to save it.
      tags:
      - Customer
      consumes:
      - "application/json"
      produces:
      - "application/csv"
      parameters:
      - name: "id"
        in: "path"
        description: "Customer id for intents"
        required: true
        type: string
      responses:
        "200":
          description: A CSV file
          schema:
            type: file
        "400":
          description: There was an error while downloading intents CSV!
...
definitions:
...
  Part:
    type: "object"
    properties:
      text:
        type: "string"
      parameterId:
        type: "string"
  TrainingPhrase:
    type: "object"
    properties:
      id:
        type: "string"
      parts:
        type: "array"
        items:
          $ref: '#/definitions/Part'
      repeatCount:
        type: "integer"
  Intent:
    type: "object"
    properties:
      displayName:
        type: "string"
      trainingPhrases:
        type: array
        items:
          $ref: '#/definitions/TrainingPhrase'
      description:
        type: "string"
      maskedIntentId:
        type: "string"
      fulfillments:
        type: "string"